requirements.txt
bcrypt==4.3.0
argon2-cffi==23.1.0

